rates: make operation timeframe a positive 2-decimal number
Change operation_timeframe from int to decimal across model/RDO/DTOs,
tighten validation to require a positive value (Range 0.01+), and
migration 010 sets the DB column to numeric(12,2) so values are stored
with two-decimal scale. The rates view now uses step/min 0.01 inputs
and formats the value with two decimals.

+++ DebantErp/DAL/Models/ProductionRateModel.cs
@@ -8,7 +8,7 @@ namespace DebantErp.DAL.Models
 
         public bool IsActual { get; set; } = true;
 
-        public int OperationTimeframe { get; set; }
+        public decimal OperationTimeframe { get; set; }
 
         public decimal Rate { get; set; }
